+++ frontend/src/pages/UserInfoPage.js
@@ -2,13 +2,11 @@ import { useState, useEffect, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import { useToken } from "../auth/useToken";
-import getUserFromToken from "../util/getUserFromToken";
 import { UserContext } from "../contexts/UserContextProvider";
 
 export const UserInfoPage = () => {
   const context = useContext(UserContext);
   const user = context.user;
-  //   console.log("in user info page", user);
   const [token, setToken] = useToken();
   // These states are bound to the values of the text inputs
   // on the page (see JSX below).
@@ -65,7 +63,6 @@ export const UserInfoPage = () => {
 
       const { token: newToken } = response.data;
       setToken(newToken);
-      context.updateUser(getUserFromToken(newToken));
       setShowSuccessMessage(true);
     } catch (error) {
       setShowErrorMessage(true);
@@ -74,6 +71,7 @@ export const UserInfoPage = () => {
 
   const logOut = () => {
     localStorage.removeItem("token");
+    setToken(null);
     navigate("/login");
   };
