+++ frontend/src/pages/UserLoginPage.js
@@ -4,7 +4,6 @@ import "../static/css/UserLoginPage.css";
 import axios from "axios";
 import { useToken } from "../auth/useToken";
 import { useQueryParams } from "../util/useQueryParams.js";
-import getUserFromToken from "../util/getUserFromToken";
 import { UserContext } from "../contexts/UserContextProvider";
 
 export const UserLoginPage = () => {
@@ -18,17 +17,10 @@ export const UserLoginPage = () => {
   const navigate = useNavigate();
   const { token: oauthToken } = useQueryParams();
 
-  // When token changes, sets the parent components user state.
-  // From the new token
-  useEffect(() => {
-    context.updateUser(getUserFromToken(token));
-  }, [token, context]);
-
   //   if there is an update in oauthToken sets the token to the new oauthToken
   useEffect(() => {
     if (oauthToken) {
       setToken(oauthToken);
-      context.updateUser(getUserFromToken(oauthToken));
       navigate("/user");
     }
   }, [oauthToken, context, setToken, navigate]);
